fixup: moved SystemUser to util and added tests.
